requirements.txt
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
pyserial==3.5
scipy==1.6.1
seaborn==0.11.1
Twisted==21.2.0
black==19.10b0
isort==5.6.4
flake8==3.8.4
torch==1.7.0
scikit-learn==0.22.2.post1